Additional cleanup
