+++ src/cfnlint/conditions/__init__.py
@@ -3,4 +3,6 @@ Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: MIT-0
 """
 
+__all__ = ["Conditions"]
+
 from cfnlint.conditions.conditions import Conditions
