+++ src/cfnlint/conditions/_equals.py
