+++ src/cfnlint/context/__init__.py
@@ -1,1 +1,3 @@
+__all__ = ["Context", "create_context_for_template"]
+
 from cfnlint.context.context import Context, create_context_for_template
