+++ src/cfnlint/jsonschema/__init__.py
@@ -1,5 +1,16 @@
 from cfnlint.jsonschema._resolver import RefResolver
-from cfnlint.jsonschema._typing import ValidationResult
+from cfnlint.jsonschema._typing import ResolutionResult, V, ValidationResult
 from cfnlint.jsonschema.exceptions import ValidationError
 from cfnlint.jsonschema.protocols import Validator
 from cfnlint.jsonschema.validators import CfnTemplateValidator, StandardValidator
+
+__all__ = [
+    ValidationResult,
+    ValidationError,
+    Validator,
+    CfnTemplateValidator,
+    StandardValidator,
+    RefResolver,
+    ResolutionResult,
+    V,
+]
