+++ src/cfnlint/rules/functions/ForEach.py
@@ -5,8 +5,7 @@ SPDX-License-Identifier: MIT-0
 
 import logging
 
-from cfnlint.languageExtensions import LanguageExtensions
-from cfnlint.rules import CloudFormationLintRule
+from cfnlint.rules import CloudFormationLintRule, RuleMatch
 
 LOGGER = logging.getLogger("cfnlint")
 
@@ -18,6 +17,20 @@ class ForEach(CloudFormationLintRule):
     source_url = "https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-getatt.html"
     tags = ["functions", "foreach"]
 
+    def validate_transform_is_declared(
+        self, has_language_extensions_transform, tree, intrinsic_function
+    ):
+        matches = []
+        if not has_language_extensions_transform:
+            message = (
+                "Missing Transform: Declare the AWS::LanguageExtensions Transform"
+                " globally to enable use of the intrinsic function "
+                + intrinsic_function
+                + " at {0}"
+            )
+            matches.append(RuleMatch(tree[:], message.format("/".join(map(str, tree)))))
+        return matches
+
     # pylint: disable=unused-argument
     def match(self, cfn):
         matches = []
@@ -27,12 +40,12 @@ class ForEach(CloudFormationLintRule):
         for for_each in for_eaches:
             has_language_extensions_transform = cfn.has_language_extensions_transform()
 
-            LanguageExtensions.validate_transform_is_declared(
-                self,
-                has_language_extensions_transform,
-                matches,
-                for_each[:-1],
-                intrinsic_function,
+            matches.extend(
+                self.validate_transform_is_declared(
+                    has_language_extensions_transform,
+                    for_each[:-1],
+                    intrinsic_function,
+                )
             )
 
         return matches
