+++ src/cfnlint/rules/functions/ToJsonString.py
@@ -7,7 +7,6 @@ from typing import Any, Dict
 
 from cfnlint.helpers import PSEUDOPARAMS
 from cfnlint.jsonschema import ValidationError, ValidationResult, Validator
-from cfnlint.languageExtensions import LanguageExtensions
 from cfnlint.rules.functions._BaseFn import BaseFn
 
 
@@ -34,37 +33,6 @@ class ToJsonString(BaseFn):
                 "Ref",
             ),
         )
-
-    def match(self, cfn):
-        has_language_extensions_transform = cfn.has_language_extensions_transform()
-        unsupported_pseudo_parameters = ["AWS::NotificationARNs"]
-
-        matches = []
-        intrinsic_function = "Fn::ToJsonString"
-        fn_toJsonString_objects = cfn.search_deep_keys(intrinsic_function)
-
-        for fn_toJsonString_object in fn_toJsonString_objects:
-            tree = fn_toJsonString_object[:-1]
-            fn_toJsonString_object_value = fn_toJsonString_object[-1]
-            LanguageExtensions.validate_transform_is_declared(
-                self,
-                has_language_extensions_transform,
-                matches,
-                tree,
-                intrinsic_function,
-            )
-            LanguageExtensions.validate_type(
-                self, fn_toJsonString_object_value, matches, tree, intrinsic_function
-            )
-            LanguageExtensions.validate_pseudo_parameters(
-                self,
-                fn_toJsonString_object_value,
-                matches,
-                tree,
-                unsupported_pseudo_parameters,
-                intrinsic_function,
-            )
-        return matches
 
     def schema(self, validator: Validator, instance: Any) -> Dict[str, Any]:
         return {
@@ -93,7 +61,10 @@ class ToJsonString(BaseFn):
                     "'AWS::LanguageExtensions' transform"
                 ),
                 validator=self.fn.py,
+                rule=self,
             )
             return
 
-        yield from super().validate(validator, s, instance, schema)
+        for err in super().validate(validator, s, instance, schema):
+            err.rule = self
+            yield err
