+++ src/cfnlint/schema/__init__.py
@@ -1,4 +1,15 @@
-from cfnlint.schema.exceptions import ResourceNotFoundError
-from cfnlint.schema.getatts import AttributeDict, GetAtt, GetAttType
+__all__ = [
+    "ResourceNotFoundError",
+    "AttributeDict",
+    "GetAtt",
+    "GetAttType",
+    "PROVIDER_SCHEMA_MANAGER",
+    "Schema",
+    "SchemaPatch",
+]
+
+from cfnlint.schema._exceptions import ResourceNotFoundError
+from cfnlint.schema._getatts import AttributeDict, GetAtt, GetAttType
+from cfnlint.schema._schema import Schema
 from cfnlint.schema.manager import PROVIDER_SCHEMA_MANAGER
-from cfnlint.schema.schema import Schema
+from cfnlint.schema.patch import SchemaPatch
