+++ src/cfnlint/schema/_exceptions.py
