+++ src/cfnlint/schema/_getatts.py
