--- test/fixtures/templates/bad/functions/toJsonStringWithoutTransform.yaml
@@ -1,10 +1,0 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Resources:
-  Secret:
-    Type: AWS::SecretsManager::Secret
-    Properties:
-      Name: TopSecret
-      SecretString:
-        Fn::ToJsonString:
-          password: "Pa$$word"
-          secretToken: "123"
