+++ test/unit/rules/functions/test_length.py
@@ -51,6 +51,7 @@ def context(cfn):
                     path=deque([]),
                     schema_path=deque([]),
                     validator="fn_length",
+                    rule=Length(),
                 ),
             ],
         ),
@@ -72,6 +73,7 @@ def context(cfn):
                     path=deque(["Fn::Length"]),
                     schema_path=deque(["type"]),
                     validator="fn_length",
+                    rule=Length(),
                 ),
             ],
         ),
@@ -86,6 +88,7 @@ def context(cfn):
                     path=deque([]),
                     schema_path=deque([]),
                     validator="fn_length",
+                    rule=Length(),
                 ),
             ],
         ),
