+++ test/unit/rules/functions/test_tojsonstring.py
@@ -52,6 +52,7 @@ def context(cfn):
                     path=deque([]),
                     schema_path=deque([]),
                     validator="fn_tojsonstring",
+                    rule=ToJsonString(),
                 ),
             ],
         ),
@@ -66,6 +67,7 @@ def context(cfn):
                     path=deque([]),
                     schema_path=deque([]),
                     validator="fn_tojsonstring",
+                    rule=ToJsonString(),
                 ),
             ],
         ),
@@ -80,6 +82,7 @@ def context(cfn):
                     path=deque(["Fn::ToJsonString"]),
                     schema_path=deque(["minProperties"]),
                     validator="fn_tojsonstring",
+                    rule=ToJsonString(),
                 ),
             ],
         ),
@@ -94,6 +97,7 @@ def context(cfn):
                     path=deque(["Fn::ToJsonString"]),
                     schema_path=deque(["minItems"]),
                     validator="fn_tojsonstring",
+                    rule=ToJsonString(),
                 ),
             ],
         ),
@@ -113,6 +117,7 @@ def context(cfn):
                     path=deque(["Fn::ToJsonString", "Ref"]),
                     schema_path=deque(["ref", "enum"]),
                     validator="ref",
+                    rule=ToJsonString(),
                 ),
             ],
         ),
